MessageApi

+++ API/UserApi/Helpers/PagedList.cs
@@ -4,7 +4,7 @@ namespace UserAPI.Helpers;
 
 public class PagedList<T> : List<T>
 {
-    public PagedList(IEnumerable<T> items, int count, int pageNumber, int pageSize)
+    public PagedList(IEnumerable<T> items, long count, int pageNumber, int pageSize)
     {
         CurrentPage = pageNumber;
         TotalPages = (int)Math.Ceiling(count / (double)pageSize);
@@ -16,7 +16,7 @@ public class PagedList<T> : List<T>
     public int CurrentPage { get; set; }
     public int TotalPages { get; set; }
     public int PageSize { get; set; }
-    public int TotalCount { get; set; }
+    public long TotalCount { get; set; }
 
     public static async Task<PagedList<T>> CreateAsync(IQueryable<T> source, int pageNumber, int pageSize)
     {
