+++ API/UserApi/Helpers/PaginationHeader.cs
@@ -2,7 +2,7 @@
 
 public class PaginationHeader
 {
-    public PaginationHeader(int currentPage, int itemsPerPage, int totalItems, int totalPages)
+    public PaginationHeader(int currentPage, int itemsPerPage, long totalItems, int totalPages)
     {
         CurrentPage = currentPage;
         ItemsPerPage = itemsPerPage;
@@ -12,6 +12,6 @@ public class PaginationHeader
 
     public int CurrentPage { get; set; }
     public int ItemsPerPage { get; set; }
-    public int TotalItems { get; set; }
+    public long TotalItems { get; set; }
     public int TotalPages { get; set; }
 }
